requirements.txt
Flask
Flask-Cors
python-dotenv
gunicorn
pdf2image
Pillow
google-generativeai